requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
requests==2.31.0
python-dotenv==1.0.0
django-allauth==0.60.1
dj-rest-auth==5.0.2
